+++ web/api/urls.py
@@ -9,6 +9,8 @@ router = routers.DefaultRouter()
 
 router.register(r'listDatatableSubdomain', SubdomainDatatableViewSet)
 
+router.register(r'listTargets', ListTargetsDatatableViewSet)
+
 router.register(r'listSubdomains', SubdomainsViewSet)
 
 router.register(r'listEndpoints', EndPointViewSet)
@@ -49,6 +51,10 @@ urlpatterns = [
         'queryIps/',
         ListIPs.as_view(),
         name='listIPs'),
+    path(
+        'queryInterestingSubdomains/',
+        QueryInterestingSubdomains.as_view(),
+        name='queryInterestingSubdomains'),
     path(
         'querySubdomains/',
         ListSubdomains.as_view(),