+++ web/targetApp/models.py
@@ -225,4 +225,4 @@ class Domain(models.Model):
 	          return obj[0].id
 
     def __str__(self):
-        return self.name
+        return str(self.name)